Scoped db package for inserts only. Removed queries.
Updated documentation on config files.
Removed uncessary tests for queries.
